Feat: Se agrega la sección encargada de mostrar los videos de los dispositivo moviles, si es que tienen.
